+++ CHANGES.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Change default cooldown period to `1.0` (was `2.0`),
+  and cooldown threshold to `100` (was `1`)
+  [#195](https://github.com/ixti/sidekiq-throttled/pull/195).
+
 ### Removed
 
 - Drop Sidekiq < 7 support